==> 기업가치/part2/04. 상관분석/02. 재무지표2/01. code/C3/ITS/3. corr_y_x_compare_allcompany.ipynb <==
import pandas as pd
import numpy as np
import os
from scipy.cluster import hierarchy as hc
import matplotlib
import matplotlib.pyplot as plt
import warnings
import seaborn as sns
from scipy import stats
import math

%matplotlib inline
warnings.filterwarnings(action='ignore')

matplotlib.rcParams['font.family'] ='Malgun Gothic'
matplotlib.rcParams['axes.unicode_minus'] =False

# Load data

dir_root = '../../../../../'
current_dir = os.getcwd()
dir_master = os.path.join(dir_root, '01. Data','00. master_data')
dir_processed = os.path.join(dir_root, '01. Data','02. 재무지표2', '01. preprocessed')
dir_result = os.path.join(dir_root, '04. 상관분석','02. 재무지표2','02. result')

y = '투자비 대비 매출 (Rev/GI)s'

# 탐색기에서 파일명 만들 때, "/"사용 불가하여 처리하는 부분
y_file_name = y.replace("/","")
c1 = 'Accenture'
c2 = 'Infosys'
c3 = 'TCS'
c4 = 'Cognizant'

outperform = [c1, c2, c3]
lowperform = [c4]

# dir_result_all = os.path.join(dir_result, 'All_ITS', y)
dir_result_all = os.path.join(dir_result, 'All_ITS', y_file_name)

if os.path.exists(dir_result_all)==False:
    os.makedirs(dir_result_all)

# total 기간에 대해서, 존재하는 변수 기준으로 inner join

df_c1 = pd.read_csv(os.path.join(dir_result, c1, y_file_name, y_file_name+'.csv')).iloc[:,:5]
df_c2 = pd.read_csv(os.path.join(dir_result, c2, y_file_name, y_file_name+'.csv')).iloc[:,:5]
df_c3 = pd.read_csv(os.path.join(dir_result, c3, y_file_name, y_file_name+'.csv')).iloc[:,:5]
df_c4 = pd.read_csv(os.path.join(dir_result, c4, y_file_name, y_file_name+'.csv')).iloc[:,:5]

df_c1.rename({'corr' : df_c1.company.unique()[0]}, axis=1, inplace=True)
df_c2.rename({'corr' : df_c2.company.unique()[0]}, axis=1, inplace=True)
df_c3.rename({'corr' : df_c3.company.unique()[0]}, axis=1, inplace=True)
df_c4.rename({'corr' : df_c4.company.unique()[0]}, axis=1, inplace=True)

df_c1.drop(['company'], axis=1, inplace=True)
df_c2.drop(['company'], axis=1, inplace=True)
df_c3.drop(['company'], axis=1, inplace=True)
df_c4.drop(['company'], axis=1, inplace=True)

df_all_variable = pd.merge(pd.merge(pd.merge(df_c1, df_c2, on=['variable','group', 'importance'], how='inner'), 
                          df_c3, on=['variable','group', 'importance'], how='inner'), 
                 df_c4, on=['variable','group', 'importance'], how='inner')

df_all_variable = df_all_variable[['variable', 'group', 'importance']+outperform+lowperform]

# 0.7 이상 상관계수가 하나라도 있으면, 남기도록 outer join

df_c1 = pd.read_csv(os.path.join(dir_result, c1, y_file_name, 'compare_period_corr_07.csv'))
df_c2 = pd.read_csv(os.path.join(dir_result, c2, y_file_name, 'compare_period_corr_07.csv'))
df_c3 = pd.read_csv(os.path.join(dir_result, c3, y_file_name, 'compare_period_corr_07.csv'))
df_c4 = pd.read_csv(os.path.join(dir_result, c4, y_file_name, 'compare_period_corr_07.csv'))

# df_c1.rename({'corr_total' : df_c1.company.unique()[0]}, axis=1, inplace=True)
df_c2.rename({'corr_total' : df_c2.company.unique()[0]}, axis=1, inplace=True)
df_c3.rename({'corr_total' : df_c3.company.unique()[0]}, axis=1, inplace=True)
df_c4.rename({'corr_total' : df_c4.company.unique()[0]}, axis=1, inplace=True)

# df_c1.drop(['corr_sub', 'company'], axis=1, inplace=True)
df_c2.drop(['corr_sub', 'company'], axis=1, inplace=True)
df_c3.drop(['corr_sub', 'company'], axis=1, inplace=True)
df_c4.drop(['corr_sub', 'company'], axis=1, inplace=True)

# df_c1 = df_c1.dropna()
df_c2 = df_c2.dropna()
df_c3 = df_c3.dropna()
df_c4 = df_c4.dropna()

# df_all = pd.merge(pd.merge(df_c1, df_c3, on=['variable','group', 'importance'], how='outer'), 
#                           df_c4, on=['variable','group', 'importance'], how='outer')
                 

df_all = pd.merge(pd.merge(pd.merge(df_c1, df_c2, on=['variable','group', 'importance'], how='outer'), 
                          df_c3, on=['variable','group', 'importance'], how='outer'), 
                 df_c4, on=['variable','group', 'importance'], how='outer')
# df_all = pd.merge(pd.merge(df_c2, df_c3, on=['variable','group', 'importance'], how='outer'), 
#                           df_c4, on=['variable','group', 'importance'], how='outer')

df_all = df_all[['variable', 'group', 'importance']+list(set(outperform)&set(df_all.columns))+list(set(lowperform)&set(df_all.columns))]

# df_all = df_all[['variable', 'group', 'importance']+[c1, c3, c4]]

# 위의 두 개 inner join

df_all_merged = pd.merge(df_all, df_all_variable, how='inner', on=['variable','group','importance'], suffixes = ("_filtered", "_show"))

# df_all_merged.to_csv(os.path.join(dir_result, 'All_ITS', 'corr_y_x_over7.csv'), index=False, encoding='utf-8-sig')

writer = pd.ExcelWriter(os.path.join(dir_result_all, 'corr_y_x_over7.xlsx'), engine='xlsxwriter')
df_all_merged.to_excel(writer, index=False)
writer.save()

df_all_merged.columns

df_all_merged
